add handleApiCall to redux-common exports. Move stuff around to avoid dependency cycles.
