+++ packages/veritone-redux-common/src/helpers/redux/fetchingStatus.js
@@ -1,0 +1,8 @@
+const fetchingStatus = {
+  fetching: 'FETCHING',
+  failure: 'FAILURE',
+  success: 'SUCCESS',
+  default: 'DEFAULT' // no action yet taken
+};
+
+export default fetchingStatus;
