+++ packages/veritone-redux-common/src/helpers/redux/paginate.test.js
@@ -1,4 +1,4 @@
-import { fetchingStatus } from 'helpers/redux';
+import fetchingStatus from 'helpers/redux/fetchingStatus';
 import paginate, {
   makeSelectors,
   clearPagination
